+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.23.2
 
 * minor changes to DictObject's documentation
+* added methods reset_and_stop and property stopped to measure